agents/requirements.txt
anthropic>=0.40.0
supabase>=2.10.0
httpx>=0.27.0
python-dotenv>=1.0.0
facebook-business>=19.0.0
google-ads>=24.0.0
google-protobuf>=4.0.0